Refactored and reorganized some code
- Changed `pages` dir to `components`. `pages` sounds like it's HTML, when
  really this package contains React components.
- Formatted some code to 2 spaces. We will adopt some JavaScript coding
  standards for the project - 2 spaces is commonly used for frontend development
  because of the deep nesting that is achieved with JSX.

Related work items: #38
