+++ src/main/app/src/App.js
@@ -1,10 +1,10 @@
 import NavBar from './components/NavBar'
 
 function App() {
-  return(
-      <div>
-         <NavBar/>
-      </div>
+  return (
+    <div>
+      <NavBar />
+    </div>
   );
 }
