+++ src/main/app/src/components/About.js
@@ -1,0 +1,9 @@
+function About() {
+  return (
+    <div>
+      <h1>This is the About us page</h1>
+    </div>
+  );
+}
+
+export default About;
