+++ src/main/app/src/components/Accounting.js
@@ -1,0 +1,9 @@
+function Accounting() {
+  return (
+    <div>
+      <h1>This is the Accounting page</h1>
+    </div>
+  );
+}
+
+export default Accounting;
