+++ src/main/app/src/components/Help.js
@@ -1,0 +1,9 @@
+function Help() {
+  return (
+    <div>
+      <h1>This is the  Help page</h1>
+    </div>
+  );
+}
+
+export default Help;
