+++ src/main/app/src/components/Home.js
@@ -1,0 +1,9 @@
+function Home() {
+  return (
+    <div>
+      <h1>This is the Home page</h1>
+    </div>
+  );
+}
+
+export default Home;
