+++ src/main/app/src/components/NavBar.js
@@ -2,7 +2,7 @@ import { AppBar, CssBaseline, Drawer, IconButton, List, ListItemIcon, Toolbar } 
 import InputBase from '@material-ui/core/InputBase';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
-import { fade, makeStyles, useTheme } from "@material-ui/core/styles";
+import { useTheme } from "@material-ui/core/styles";
 import Typography from '@material-ui/core/Typography';
 import { ChevronLeft } from '@material-ui/icons';
 import AccountBalanceIcon from '@material-ui/icons/AccountBalance';
@@ -22,117 +22,17 @@ import SearchIcon from '@material-ui/icons/Search';
 import clsx from 'clsx';
 import React, { useState } from 'react';
 import { BrowserRouter as Router, Link, Route, Switch } from 'react-router-dom';
-import About from '../pages/About';
-import Accounting from '../pages/Accounting';
-import Help from '../pages/Help';
-import Home from '../pages/Home';
-import { Inventory } from '../pages/Inventory';
-import Planning from '../pages/Planning';
-import Production from '../pages/Production';
-import Purchase from '../pages/Purchase';
-import Sales from '../pages/Sales';
-import UserAccount from '../pages/UserAccount';
-
-const drawerWidth = 200;
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    display: 'flex',
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  appBarShift: {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  menuButton: {
-    marginRight: 36,
-  },
-  hide: {
-    display: 'none',
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-    whiteSpace: 'nowrap',
-  },
-  drawerOpen: {
-    width: drawerWidth,
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  drawerClose: {
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    overflowX: 'hidden',
-    width: theme.spacing(6) + 1,
-    [theme.breakpoints.up('sm')]: {
-      width: theme.spacing(7) + 3,
-    }
-  },
-  toolbar: {
-    backgroundColor: theme.palette.primary,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-  },
-  link: {
-    textDecoration: 'none',
-    color: theme.palette.text.primary
-  },
-  search: {
-    position: 'relative',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.white, 0.15),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white, 0.25),
-    },
-    margin: 'auto',
-    width: '44%',
-  },
-  searchIcon: {
-    padding: theme.spacing(0, 2),
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  inputRoot: {
-    color: 'inherit',
-  },
-  inputInput: {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('md')]: {
-      width: '20ch',
-    },
-  },
-}));
-
+import About from './About';
+import Accounting from './Accounting';
+import Help from './Help';
+import Home from './Home';
+import Planning from './Planning';
+import Production from './Production';
+import Purchase from './Purchasing';
+import Sales from './Sales';
+import UserAccount from './UserAccount';
+import useNavBarStyles from '../styles/NavBarSyles';
+import { Inventory } from './inventory/Inventory';
 
 const DrawerItem = ({ cls, Icon, link, text }) => {
   return (
@@ -167,6 +67,7 @@ const DrawerList = ({ drawerItemClass }) => {
             cls={drawerItemClass}
             link={path}
             text={text}
+            key={text}
           />
         ))
       }
@@ -174,9 +75,26 @@ const DrawerList = ({ drawerItemClass }) => {
   );
 };
 
+const ContentSwitch = () => {
+  return (
+    <Switch>
+      <Route exact path="/" component={Home} />
+      <Route exact path="/planning" component={Planning} />
+      <Route exact path="/inventory" component={Inventory} />
+      <Route exact path="/production" component={Production} />
+      <Route exact path="/accounting" component={Accounting} />
+      <Route exact path="/sales" component={Sales} />
+      <Route exact path="/purchasing" component={Purchase} />
+      <Route exact path="/help" component={Help} />
+      <Route exact path="/about" component={About} />
+      <Route exact path="/useraccount" component={UserAccount} />
+    </Switch>
+  );
+};
+
 function NavBar() {
 
-  const classes = useStyles();
+  const classes = useNavBarStyles();
   const theme = useTheme();
   const [open, setOpen] = useState(false);
 
@@ -253,18 +171,7 @@ function NavBar() {
         </Drawer>
         <main className={classes.content} >
           <div className={classes.toolbar} />
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <Route exact path="/planning" component={Planning} />
-            <Route exact path="/inventory" component={Inventory} />
-            <Route exact path="/production" component={Production} />
-            <Route exact path="/accounting" component={Accounting} />
-            <Route exact path="/sales" component={Sales} />
-            <Route exact path="/purchasing" component={Purchase} />
-            <Route exact path="/help" component={Help} />
-            <Route exact path="/about" component={About} />
-            <Route exact path="/useraccount" component={UserAccount} />
-          </Switch>
+          <ContentSwitch />
         </main>
       </Router>
     </div>
