+++ src/main/app/src/components/Planning.js
@@ -1,0 +1,10 @@
+
+function Planning() {
+  return (
+    <div>
+      <h1>This is the Planning page</h1>
+    </div>
+  );
+}
+
+export default Planning;
