+++ src/main/app/src/components/Production.js
@@ -1,0 +1,9 @@
+function Production() {
+  return (
+    <div>
+      <h1>This is the Production page</h1>
+    </div>
+  );
+}
+
+export default Production;
