+++ src/main/app/src/components/Sales.js
@@ -1,0 +1,9 @@
+function Sales() {
+  return (
+    <div>
+      <h1>This is the Sales page</h1>
+    </div>
+  );
+}
+
+export default Sales;
