+++ src/main/app/src/components/UserAccount.js
@@ -1,0 +1,9 @@
+function UserAccount() {
+  return (
+    <div>
+      <h1>This is the User Account page</h1>
+    </div>
+  );
+}
+
+export default UserAccount;
