+++ src/main/app/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './styles/index.css';
 import App from './App'
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './misc/reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
