+++ src/main/app/src/misc/React-Spinner.css
