--- src/main/app/src/pages/About.js
@@ -1,10 +1,0 @@
-function About() {
-    return(
-        <div>
-            <h1>This is the About us page</h1>
-        </div>
-    );
-    
-}
-
-export default About;
