--- src/main/app/src/pages/Accounting.js
@@ -1,10 +1,0 @@
-function Accounting() {
-    return(
-        <div>
-            <h1>This is the Accounting page</h1>
-        </div>
-    );
-    
-}
-
-export default Accounting;
