--- src/main/app/src/pages/Help.js
@@ -1,10 +1,0 @@
-function Help() {
-    return(
-        <div>
-            <h1>This is the  Help page</h1>
-        </div>
-    );
-    
-}
-
-export default  Help;
