--- src/main/app/src/pages/Home.js
@@ -1,12 +1,0 @@
-
-
-function Home() {
-    return(
-        <div>
-           <h1>This is the Home page</h1>
-        </div>
-    );
-    
-}
-
-export default Home;
