--- src/main/app/src/pages/Planning.js
@@ -1,11 +1,0 @@
-
-function Planning() {
-    return(
-        <div>
-            <h1>This is the Planning page</h1>
-        </div>
-    );
-    
-}
-
-export default Planning;
