--- src/main/app/src/pages/Production.js
@@ -1,10 +1,0 @@
-function Production() {
-    return(
-        <div>
-            <h1>This is the Production page</h1>
-        </div>
-    );
-    
-}
-
-export default Production;
