--- src/main/app/src/pages/Sales.js
@@ -1,10 +1,0 @@
-function Sales() {
-    return(
-        <div>
-            <h1>This is the Sales page</h1>
-        </div>
-    );
-    
-}
-
-export default Sales;
