--- src/main/app/src/pages/UserAccount.js
@@ -1,10 +1,0 @@
-function UserAccount() {
-    return(
-        <div>
-            <h1>This is the User Account page</h1>
-        </div>
-    );
-    
-}
-
-export default  UserAccount;
